requirements-dev.txt
alabaster==0.7.12
astroid==2.8.0
attrs==21.2.0
autoflake==1.4
Babel==2.9.1
backports.entry-points-selectable==1.1.0
black==21.9b0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
codespell==2.1.0
colorama==0.4.4
coverage==5.5
cryptography==3.4.8
distlib==0.3.2
docutils==0.17.1
filelock==3.0.12
flake8==3.9.2
idna==3.2
imagesize==1.2.0
importlib-metadata==4.8.1
iniconfig==1.1.1
isort==5.9.3
jeepney==0.7.1
Jinja2==3.0.1
jsonpointer==2.1
keyring==23.2.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pkginfo==1.7.1
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pydantic-yaml==0.4.2
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.10.0
pylint==2.12.2
pylint-fixme-info==1.0.3
pylint-pytest==1.1.2
pyparsing==2.4.7
pytest==6.2.5
pytest-mock==3.6.1
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
readme-renderer==29.0
regex==2021.8.28
requests==2.26.0
requests-mock==1.9.3
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==3.3.1
semver==2.13.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-autodoc-typehints==1.12.0
sphinx-jsonschema==1.16.11
sphinx-pydantic==0.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==1.2.1
tox==3.24.4
tqdm==4.62.3
twine==3.4.2
types-PyYAML==5.4.10
types-requests==2.25.7
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.8.0
webencodings==0.5.1
wrapt==1.12.1
zipp==3.5.0
